requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
fastapi==0.88.0
    # via -r requirements.in
grpcio==1.51.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.51.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.2
    # via httpx
httptools==0.5.0
    # via -r requirements.in
httpx==0.23.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
protobuf==4.21.9
    # via
    #   -r requirements.in
    #   grpcio-tools
pydantic==1.10.2
    # via fastapi
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.22.0
    # via fastapi
typing-extensions==4.4.0
    # via pydantic
uvicorn==0.20.0
    # via -r requirements.in
uvloop==0.17.0
    # via -r requirements.in
